+++ backend/models/schemas/user.schema.ts
@@ -1,4 +1,5 @@
 import { object, string, TypeOf, nativeEnum, array } from "zod";
+import { postSchema } from "./post.schema";
 import { UserType } from "../user.models";
 // ensure that the incoming data is validated and in correct format
 export const registerUserSchema = object({
@@ -14,12 +15,29 @@ export const registerUserSchema = object({
       required_error: "Password confirmation is required",
     }).trim(),
     role: nativeEnum(UserType, { required_error: "Role is required" }),
-    // ! TODO : create a schema for Post and pass it into array parameter
-    // pins: array().optional(),
   }).refine((data) => data.password === data.passwordConfirmation, {
     message: "Passwords do not match",
     path: ["passwordConfirmation"],
   }),
 });
+
+registerUserSchema.extend({ pins: array(postSchema).optional() });
+
+export const loginUserSchema = object({
+  body: object({
+    email: string({
+      required_error: "Email is required",
+    })
+      .trim()
+      .email("Invalid email or password"),
+    password: string({
+      required_error: "Password is required",
+    })
+      .trim()
+      .min(8, "Invalid email or password"),
+  }),
+});
+
+export type LoginUserInput = TypeOf<typeof loginUserSchema>["body"];
 
 export type RegisterUserInput = TypeOf<typeof registerUserSchema>["body"];
